Disable throttling everywhere
Last time I only disabled throttling in one file, assuming, that
this would be the central place to do so. This is not the case, so I
disabled it in the remaining code areas.
I also had to reduce the number of called nock mock calls for issue
comments in the tests. The comment said it was necessary, but that
might be outdated and disabling thorttling reduced the number of calls.