Remove snapshot match from oauth revocation
I noticed, that one of the most flaky tests is the snapshot match for oauth verification.
See [this bug tracking issue](https://github.com/integrations/slack/issues/903#issuecomment-519888747) for details.
The core problem is that we are combining `machtesSnapshot` with `times(x)`.
Combining these requires a ordered or sequential execution of the requests
that are performed. But for `signout` those are executed in parallel,
making the order unpredictable which causes snapshot mismatches.
See https://github.com/integrations/slack/blob/b5cf9306a26e93bcf0be59cd84eda97865566e33/lib/github/oauth.js#L128-L140 for the implementation.

This PR removes the snapshot match, which is not reliable and is not
as important as the remaining `expect` clauses that validate the actual
removal.