Improve integration test stability (#906)
* Makes unfurl snapshot test deterministic

Concurrent unfurl bodies are tested in a non depterministic way, by
running it in combination with `times`. This In most cases the
first unfurl will be a little bit earlier, so that in more than 50% of
the time, the current snapshot will match. However, if the second
request happens earlier (depends on scheduler timing), both snapshots
comparisons will fail, because the first is compared to the second and
visa versa. For a full explanation see: https://github.com/integrations/slack/issues/903#issuecomment-520333620

Refs: #903

* Make slack-uninstall snapshot tests reliable

By adding a order statement to `findAll`, I made the order determinstic,
to fix test failures.
For a full explanation see: https://github.com/integrations/slack/issues/903#issuecomment-520333620

Refs #903