Fix request sorting in unfurl tests and add sleep
We are experiencing a lot of test failures to due missing mocks in
the notification tests. I added a sleep and `activeMock` assertions
to get more insights. These started happening after the Probot update.
In addition the unfor request snapshot was not properly sorted.
Now they are sorted by the unfurl URL per request. Hope that works.