Disable flaky integration test
The issue-comment test is failing almost all the time on CI.
I decided to add it to this PR instead of a dedicated PR, because
it is blocking shipping this feature.
For tracking the problem and to make sure it will be enabled again,
refer to this issue: https://github.com/integrations/slack/issues/903#issuecomment-521535017

I also fixed the way we sort entries in the unfurl test.