Make unsubscribe work again
After hacking on the internal APIs for a while, this commit fixes the
end2end unsubscribe feature.
This is still WIP, but ow have a working unsubscribe and subscribe again.

Bad news: Due to the current API mismatch between parsing and
subscribing, more than 50 are failing.
I need to figure out the abstactions I want to have, to address this
in a backward-compatible way.