Handle invalid subscribe labels
If an invalid subscribe label is entered, the user receives a nice
error message showing the invalid input.
In addition I made sure, that we tread (un)subscribe command that only
contain invalid entries as entries with arguments.
This helps us to avoid a tricky issue with unsubscibe:

If only an invalid label is given and `hasValue` would be false, we would call unsubscribe instead of disable,
causing the entire subscription to be deleted, even though the user intended to remove a certain label.