Render labels nicely when listing features
Previously the list of rendered labels looked like this:
`label:label1, labelb`.
This becomes unreadable as soon as text emojis or spaces are used.
For example: `label::construction: WIP` is not great at all.

After this change arrays are rendered like this:
`label: [':construction: WIP', 'DO NOT MERGE']`
which is much more readable.