Use deployment-status ref instead of sha
When the integrator that created the deployment status provided a ref,
that is different from the commit sha, that ref is now displayed instead
of the short sha.
This addresses the issue discussed in #98 and #85.
I think the next step would be doing an actual API request to fetch the
branches, but his change is a 0-cost improvement and (almost) 0-effort
improvement that allows us to gather metrics about custom ref adoption
and maybe this already solves the problem.

Refs #85, #98