Implement new filter syntax.
After great feedback, I decided to change the filter syntax to avoid
name conflicts with the GitHub `label` webhook.
The new syntax is:
`+label:myLabel` for requiring a label and in the future
`-label:myLabel` for ignoring a label.

I will update the docs and do cleanup in the next commit.