📘 Improve docs about filter-affected events
Since it is not intuitive which events are affected by the
`required-label` filter, I added an overview to the README.
Right now it is quite simple, because everything that is related to a
label is filtered, but this doesn't mean we implemented it that way. In
order to avoid questions and confusions, it makes sense to make the docs
very explict. Especially since this will be really important once we
have more than one filter.