Improve logging in error-handler
We received multiple error reports that did not contain a sentry
reference and if it did, the information we were able was not sufficient
to debug this issue.
This change addresses the issue by logging more details in case of an internal error that is
reported to the user via an ephemeral message.