Address review concerns and simplify
I decided to leave `res.sentry` the way it was and pass a ref that
always exists to all the functions that need a reference (for the user).
The risk should be small because I did not find other places consuming
`res.sentry`. In addition the likelihood of making the mistake of
`hasSentryRef` being always true, again, is eliminated.

Finally I adjusted some field names and added a `gh-` prefix to our
customRef to make sure support can distinguish sentry and loggly refs.